requirements.txt
torch
datasets
trl==0.15.2
qwen_vl_utils
mathruler
pylatexenc
Pillow
vllm>0.7.2
torchvision

# easyr1
omegaconf
tensordict
torchdata
tensorboard
codetiming
pebble
timeout_decorator
word2number
sentence_transformers
